Move dir to ex 20, use explicit events and dependencies and cosntexpr func to calculate min
