+++ Code_Exercises/Exercise_20_SYCL_Reduction/Makefile
